Add files via upload
ASSIGNMENT#3

Exercise 1:
Write a Python program to create a list of 5 integers and perform the following
operations:
○ Print the list.
○ Append a new integer to the list.
○ Sort the list in ascending order.
○ Reverse the elements of the list.

Exercise 2:
Create a tuple of 4 elements and perform the following operations:
○ Print the tuple.
○ Access the third element of the tuple.
○ Concatenate the tuple with another tuple of 2 elements.
○ Print the length of the concatenated tuple.

Exercise 3:
Create a dictionary representing a student record with keys: 'name', 'roll_number',
'marks'. Perform the following operations:
○ Print the dictionary.
○ Update the marks to 85.
○ Check if 'grade' key exists, if not, add 'grade' with value 'A'.

Exercise 4:
Write a Python program that takes two lists of integers as input and returns a new list
containing elements common to both lists, without duplicates.
Exercise 5:
Write a program that reads a list of integers and returns a new list containing only the
even numbers using list.
Exercise 6:
Write a Python function that takes a list of tuples, each tuple containing a student's
name and their average score, and returns a list of names sorted by their scores in
descending order.
Exercise 7:

Write a program that checks if one set is a subset of another set.
Exercise 8:
Implement a function that takes a list of strings as input and returns a new list of
strings sorted by the length of each string, from shortest to longest.
Exercise 9:
Write a Python program to implement a function that takes a list of integers as input
and returns a new list containing only the unique elements, preserving their original
order.
Exercise 10:
Create a dictionary representing a word frequency counter from a given list of words.
Write a program to find the top 3 most frequent words and their frequencies.